refactor teamcodebeta
