+++ TeamCodeBeta/src/main/java/org/firstinspires/ftc/teamcodebeta/drive/roadrunner/TurnTest.java
@@ -1,0 +1,29 @@
+package org.firstinspires.ftc.teamcodebeta.drive.roadrunner;
+
+import com.acmerobotics.dashboard.config.Config;
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+import org.firstinspires.ftc.teamcodebeta.drive.config.SampleMecanumDrive;
+
+/*
+ * This is a simple routine to test turning capabilities.
+ */
+@Config
+@Autonomous(group = "drive")
+@Disabled
+public class TurnTest extends LinearOpMode {
+    public static double ANGLE = 90; // deg
+
+    @Override
+    public void runOpMode() throws InterruptedException {
+        SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
+
+        waitForStart();
+
+        if (isStopRequested()) return;
+
+        drive.turn(Math.toRadians(ANGLE));
+    }
+}
