+++ TeamCodeBeta/src/main/java/org/firstinspires/ftc/teamcodebeta/drive/samples/BasicGoBildaStraferOp.java
@@ -1,0 +1,129 @@
+package org.firstinspires.ftc.teamcodebeta.drive.samples;
+
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.hardware.Servo;
+
+@TeleOp(name = "BasicGoBildaStraferOp")
+@Disabled
+public class BasicGoBildaStraferOp extends LinearOpMode {
+
+    private DcMotor rightFrontDrive;
+    private DcMotor rightRearDrive;
+    private DcMotor leftRearDrive;
+    private DcMotor leftFrontDrive;
+    private DcMotor clawArm;
+
+    private Servo clawClose;
+    private Servo clawLift;
+
+
+    /**
+     * powerAdjust - Adjust the power based on the percent passed in.
+     * @param power - Power value to adjust
+     * @param percent - Percent to adjust power by
+     * @return - Adjusted power value
+     */
+    private double powerAdjust(float power, double percent) {
+        double result;
+
+        result = (power * percent) / 100;
+        return result;
+    }
+
+    /**
+     * This function is executed when this Op Mode is selected from the Driver Station.
+     */
+    @Override
+    public void runOpMode() {
+        int drivePercentage;
+        int strafePercentage;
+        int rotatePercentage;
+        double driveValue;
+        double strafeValue;
+        double rotateValue;
+        double clawArmPower;
+
+        rightFrontDrive = hardwareMap.get(DcMotor.class, "rightFrontDrive");
+        rightRearDrive = hardwareMap.get(DcMotor.class, "rightRearDrive");
+        leftRearDrive = hardwareMap.get(DcMotor.class, "leftRearDrive");
+        leftFrontDrive = hardwareMap.get(DcMotor.class, "leftFrontDrive");
+
+        clawArm = hardwareMap.get(DcMotor.class, "claw_arm");
+        clawLift = hardwareMap.get(Servo.class, "claw_lift");
+        clawClose = hardwareMap.get(Servo.class, "claw_close");
+
+        // Set motor directions
+        rightFrontDrive.setDirection(DcMotor.Direction.REVERSE);
+        rightRearDrive.setDirection(DcMotor.Direction.REVERSE);
+        leftRearDrive.setDirection(DcMotor.Direction.FORWARD);
+        leftFrontDrive.setDirection(DcMotor.Direction.FORWARD);
+        // Drive speed as a percent.
+        // Set to 100 for full power.
+        drivePercentage = 75;
+        // Strafe speed as a percent.
+        // Set to 100 for full power.
+        strafePercentage = 75;
+        // Rotate speed as a percent.
+        // Set to 100 for full power.
+        rotatePercentage = 45;
+
+        waitForStart();
+        if (opModeIsActive()) {
+            while (opModeIsActive()) {
+
+                // Read Values
+                driveValue = powerAdjust(-gamepad1.left_stick_y, drivePercentage);
+                strafeValue = powerAdjust(-gamepad1.right_stick_x, strafePercentage);
+                if (gamepad1.left_trigger > 0) {
+                    rotateValue = powerAdjust(-1, rotatePercentage);
+                } else if (gamepad1.right_trigger > 0) {
+                    rotateValue = powerAdjust(1, rotatePercentage);
+                } else {
+                    rotateValue = powerAdjust(0, rotatePercentage);
+                }
+                // Read the y value of the left thumbstick of gamepad2
+                clawArmPower = -gamepad2.left_stick_y;
+
+
+                // Execute Actions
+                moveRobot(driveValue, strafeValue, rotateValue);
+                clawArm.setPower(clawArmPower);
+
+                // Telemetry Data
+                telemetry.addData("drive", driveValue);
+                telemetry.addData("strafe", strafeValue);
+                telemetry.addData("rotate", rotateValue);
+                telemetry.addData("clawArmPower", clawArmPower);
+                telemetry.update();
+            }
+        }
+    }
+
+    /**
+     * moveRobot - Move the robot based on the power values passed in.
+     * @param drive - Drive power value
+     *              Positive values = forward
+     *              Negative values = backward
+     *              Range -1.0 to 1.0
+     *              0 = no drive
+     * @param strafe - Strafe power value
+     *               Positive values = right
+     *               Negative values = left
+     *               Range -1.0 to 1.0
+     *               0 = no strafe
+     * @param rotate - Rotate power value
+     *               Positive values = clockwise
+     *               Negative values = counter clockwise
+     *               Range -1.0 to 1.0
+     *               0 = no rotate
+     */
+   private void moveRobot(double drive, double strafe, double rotate) {
+        leftFrontDrive.setPower((drive + rotate) - strafe);
+        rightFrontDrive.setPower((drive - rotate) + strafe);
+        leftRearDrive.setPower(drive + rotate + strafe);
+        rightRearDrive.setPower((drive - rotate) - strafe);
+    }
+}
