+++ TeamCodeBeta/src/main/java/org/firstinspires/ftc/teamcodebeta/util/LogFiles.java
@@ -16,10 +16,10 @@ import com.qualcomm.robotcore.util.WebHandlerManager;
 
 import org.firstinspires.ftc.ftccommon.external.WebHandlerRegistrar;
 import org.firstinspires.ftc.robotcore.internal.system.AppUtil;
-import org.firstinspires.ftc.teamcodebeta.robots.base.BetaDriveConstants;
-import org.firstinspires.ftc.teamcodebeta.robots.base.SampleMecanumDrive;
-import org.firstinspires.ftc.teamcodebeta.robots.base.SampleTankDrive;
-import org.firstinspires.ftc.teamcodebeta.robots.base.StandardTrackingWheelLocalizer;
+import org.firstinspires.ftc.teamcodebeta.drive.config.BetaDriveConstants;
+import org.firstinspires.ftc.teamcodebeta.drive.config.SampleMecanumDrive;
+import org.firstinspires.ftc.teamcodebeta.drive.config.SampleTankDrive;
+import org.firstinspires.ftc.teamcodebeta.drive.config.StandardTrackingWheelLocalizer;
 
 import java.io.File;
 import java.io.FileInputStream;
